requirements.txt
xrpl-py
mysqlclient